requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
discord.py==1.2.4
firebase-admin==3.1.0
Flask==1.1.1
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
httplib2==0.14.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
kwconfig==1.1.7
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==4.5.2
protobuf==3.10.0
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-firebase==1.2
pytz==2019.3
requests==2.22.0
rsa==4.0
search-google==1.2.1
six==1.12.0
typing-extensions==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.6
websockets==6.0
Werkzeug==0.16.0
yarl==1.3.0
